+++ ArcadiaParties-SPA/src/app/department/department-selector/department-selector.component.scss
