+++ ArcadiaParties-SPA/src/app/department/department.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { StoreModule } from '@ngrx/store';
 
-import { DepartmentIdentityComponent } from './department-identity/department-identity.component';
+import { DepartmentIdentityComponent } from './department-selector/department-selector.component';
 import { MaterialModule } from '../material/material.module';
 import { departmentsFeatureStateKey, departmentReducer } from './reducers/reducer';
 import { selectedDepartmentFeatureStateKey, selectedDepartmentReducer } from './reducers/selected-department-reducer';
